+++ components/map/MapScreen.tsx
@@ -4,7 +4,7 @@ import { useCallback, useEffect, useRef, useState } from 'react'
 import { View } from 'react-native'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 
-import Map, { MapRef } from '@/components/map/Map'
+import { MapRef } from '@/components/map/Map'
 import MapPointBottomSheet from '@/components/map/MapPointBottomSheet'
 import MapZoneBottomSheet from '@/components/map/MapZoneBottomSheet'
 import IconButton from '@/components/shared/IconButton'
@@ -48,12 +48,12 @@ const MapScreen = () => {
 
   return (
     <View className="flex-1 items-stretch">
-      <Map
+      {/* <Map
         ref={mapRef}
         onZoneChange={handleZoneChange}
         onPointPress={handlePointPress}
         filters={filters}
-      />
+      /> */}
       <MapZoneBottomSheet
         ref={zoneBottomSheetRef}
         zone={selectedZone}
