+++ package.json
@@ -54,6 +54,7 @@
     "react": "18.2.0",
     "react-i18next": "~13.2.2",
     "react-native": "~0.72.5",
+    "react-native-date-picker": "^4.3.3",
     "react-native-gesture-handler": "~2.12.0",
     "react-native-mmkv": "~2.10.2",
     "react-native-pager-view": "6.2.0",
